moved chemgen to autofrag
part of Shaun's review
